+++ buession-redis/src/main/java/com/buession/redis/core/operations/HashOperations.java
@@ -471,7 +471,7 @@ public interface HashOperations extends HashCommands, RedisOperations {
 	 *
 	 * @return 执行成功返回 Status.Success，否则返回 Status.FAILURE
 	 */
-	Status hMSet(final String key, final List<KeyValue<String, ?>> data);
+	Status hMSet(final String key, final List<KeyValue<String, Object>> data);
 
 	/**
 	 * 批量将多个 field =&gt; value (域-值)对设置到哈希表 key 中
@@ -485,7 +485,7 @@ public interface HashOperations extends HashCommands, RedisOperations {
 	 *
 	 * @return 执行成功返回 Status.Success，否则返回 Status.FAILURE
 	 */
-	Status hMSet(final byte[] key, final List<KeyValue<byte[], ?>> data);
+	Status hMSet(final byte[] key, final List<KeyValue<byte[], Object>> data);
 
 	/**
 	 * 批量将多个 field =&gt; value (域-值)对设置到哈希表 key 中
@@ -501,7 +501,7 @@ public interface HashOperations extends HashCommands, RedisOperations {
 	 *
 	 * @since 3.0.0
 	 */
-	default Status hMSet(final String key, final KeyValue<String, ?>... data) {
+	default Status hMSet(final String key, final KeyValue<String, Object>... data) {
 		return hMSet(key, Arrays.asList(data));
 	}
 
@@ -519,7 +519,7 @@ public interface HashOperations extends HashCommands, RedisOperations {
 	 *
 	 * @since 3.0.0
 	 */
-	default Status hMSet(final byte[] key, final KeyValue<byte[], ?>... data) {
+	default Status hMSet(final byte[] key, final KeyValue<byte[], Object>... data) {
 		return hMSet(key, Arrays.asList(data));
 	}